refactor part 2

+++ FTHub/App/BeforeAuth/Setup/SetupPageTrainingsPerWeekView.swift
@@ -43,7 +43,7 @@ struct SetupPageTrainingsPerWeekView: View {
             
             Button(action: {
                 user[0].userDetails?.setupActivePage += 1
-                user[0].userDetails?.workoutsPerWeek = selectedDays
+                user[0].userDetails?.workoutsPerWeek = Int16(selectedDays)
                 setupController.workoutsPerWeek = selectedDays
                 setupController.activePage += 1
                 try? context.save()
